+++ web-frontend/modules/database/components/view/grid/GridViewFieldType.vue
@@ -15,71 +15,22 @@
       >
         <i class="fas fa-caret-down"></i>
       </a>
-      <Context ref="context">
-        <ul class="context-menu">
-          <li>
-            <a
-              ref="updateFieldContextLink"
-              class="grid-view-description-options"
-              @click="
-                $refs.updateFieldContext.toggle(
-                  $refs.updateFieldContextLink,
-                  'bottom',
-                  'right',
-                  0
-                )
-              "
-            >
-              <i class="context-menu-icon fas fa-fw fa-pen"></i>
-              Edit field
-            </a>
-            <UpdateFieldContext
-              ref="updateFieldContext"
-              :field="field"
-              @update="$refs.context.hide()"
-            ></UpdateFieldContext>
-          </li>
-          <li v-if="!field.primary">
-            <a @click="deleteField(field)">
-              <i class="context-menu-icon fas fa-fw fa-trash"></i>
-              Delete field
-            </a>
-          </li>
-        </ul>
-      </Context>
+      <FieldContext ref="context" :field="field"></FieldContext>
       <slot></slot>
     </div>
   </div>
 </template>
 
 <script>
-import { notifyIf } from '@baserow/modules/core/utils/error'
-import UpdateFieldContext from '@baserow/modules/database/components/field/UpdateFieldContext'
+import FieldContext from '@baserow/modules/database/components/field/FieldContext'
 
 export default {
   name: 'GridViewFieldType',
-  components: { UpdateFieldContext },
+  components: { FieldContext },
   props: {
     field: {
       type: Object,
       required: true,
-    },
-  },
-  methods: {
-    setLoading(field, value) {
-      this.$store.dispatch('field/setItemLoading', { field, value })
-    },
-    async deleteField(field) {
-      this.$refs.context.hide()
-      this.setLoading(field, true)
-
-      try {
-        await this.$store.dispatch('field/delete', field)
-      } catch (error) {
-        notifyIf(error, 'field')
-      }
-
-      this.setLoading(field, false)
     },
   },
 }