+++ web-frontend/modules/database/mixins/gridFieldInput.js
@@ -8,11 +8,11 @@ export default {
   data() {
     return {
       /**
-       * Indicates whether of the user is editing the value.
+       * Indicates whether the user is editing the value.
        */
       editing: false,
       /**
-       *  A temporary copy of the value when editing.
+       * A temporary copy of the value when editing.
        */
       copy: null,
     }